requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
websockets==15.0.1
google-genai==1.26.0
python-dotenv==1.1.1
requests==2.32.4
twelvelabs==0.4.10 